+++ pkg/tool/schema_query_validator_test.go
@@ -154,84 +154,3 @@ func TestSchemaQueryValidator_NormalizeNamespace(t *testing.T) {
 		})
 	}
 }
-
-func TestSchemaQueryValidator_InferProviderName(t *testing.T) {
-	validator := NewSchemaQueryValidator()
-
-	tests := []struct {
-		name         string
-		category     string
-		resourceType string
-		providerName string
-		expected     string
-		expectError  bool
-		errorMsg     string
-	}{
-		{
-			name:         "existing provider name preserved",
-			category:     "resource",
-			resourceType: "azurerm_resource_group",
-			providerName: "azurerm",
-			expected:     "azurerm",
-			expectError:  false,
-		},
-		{
-			name:         "infer provider name from resource type",
-			category:     "resource",
-			resourceType: "azurerm_resource_group",
-			providerName: "",
-			expected:     "azurerm",
-			expectError:  false,
-		},
-		{
-			name:         "infer provider name from data source type",
-			category:     "data",
-			resourceType: "aws_availability_zones",
-			providerName: "",
-			expected:     "aws",
-			expectError:  false,
-		},
-		{
-			name:         "provider category requires explicit provider name",
-			category:     "provider",
-			resourceType: "",
-			providerName: "",
-			expected:     "",
-			expectError:  true,
-			errorMsg:     "provider name is required when category is 'provider'",
-		},
-		{
-			name:         "function category requires explicit provider name",
-			category:     "function",
-			resourceType: "",
-			providerName: "",
-			expected:     "",
-			expectError:  true,
-			errorMsg:     "provider name is required when category is 'function'",
-		},
-		{
-			name:         "cannot infer provider name from invalid type",
-			category:     "resource",
-			resourceType: "invalidtype",
-			providerName: "",
-			expected:     "",
-			expectError:  true,
-			errorMsg:     "could not infer provider name from type 'invalidtype', please provide the 'name' parameter",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			result, err := validator.InferProviderName(tt.category, tt.resourceType, tt.providerName)
-
-			if tt.expectError {
-				require.Error(t, err)
-				assert.Contains(t, err.Error(), tt.errorMsg)
-				assert.Equal(t, tt.expected, result)
-			} else {
-				require.NoError(t, err)
-				assert.Equal(t, tt.expected, result)
-			}
-		})
-	}
-}
